requirement.txt
flask==3.1.0
pdfplumber==0.11.4
aspose-words==24.12.0
python-dotenv==1.0.1
pyttsx3==2.98
pydub==0.25.1